start-samba: don't allow attempts at invalid hostnames
Hostnames can only be hyphens and case-insensitive alphanumerics.

Nova silently converts the names it is given into valid hostnames, but
our scripts assume the hostname is the string we gave nova. So here we
convert the invalid characters first.

This of course could lead to collisions, so don't do that.

Signed-off-by: Douglas Bagnall <douglas.bagnall@catalyst.net.nz>