start-samba: option to choose or guess net-id
When there are extra networks (due to the nefarious work of heat),
nova doesn't know which one to use so it complains and refuses to do
anything. The --net-id option lets you specify a network by UUID. For
those times when you forget your favourite UUID, --net-id=auto will
select the one it thinks you probably want, using `nova net-list` and
the dirty heuristic of picking the first one it sees.

Signed-off-by: Douglas Bagnall <douglas.bagnall@catalyst.net.nz>