multi-perf-test: --test-list to use an alternate test list
The lists of tests to run is either contained in or created by a file
like $samba/selftest/perf_tests.py. If you want the tests to be run,
you need to append a properly escaped pipe, like so:

  --test-list='~/tests.py\|'

which is what you do if the file is a samba style test generator with
plantestsuite() and so forth. Alternatively you can just catch the
output that such a file generates and go:

  --test-list=~/tests.txt